#38 feat: added sales routes to admin module
