+++ lib/app/modules/sales_route_register/domain/entity/sale_entity.dart
