+++ lib/app/modules/sales_route_register/domain/repository/sales_register_repository.dart
@@ -1,5 +1,5 @@
 import 'package:appweb/app/core/error/failures.dart';
-import 'package:appweb/app/modules/sales_register/data/models/sales_model.dart';
+import 'package:appweb/app/modules/sales_route_register/data/models/sales_model.dart';
 import 'package:dartz/dartz.dart';
 
 abstract class SalesRepository {
