+++ lib/app/modules/sales_route_register/presentation/bloc/sales_events.dart
@@ -1,4 +1,4 @@
-import 'package:appweb/app/modules/sales_register/data/models/sales_model.dart';
+import 'package:appweb/app/modules/sales_route_register/data/models/sales_model.dart';
 
 abstract class SalesEvent {}
