+++ lib/app/modules/sales_route_register/presentation/bloc/sales_states.dart
