+++ lib/app/modules/sales_route_register/presentation/content/content_tablet_sales.dart
